+++ changelog.md
@@ -1,1 +1,2 @@
 * Added Molang animation support
+* Added api for player animations